change AssetId::Asset from `u64` to `u128`

+++ src/assets.rs
@@ -30,7 +30,7 @@ pub enum AssetId {
     /// Generic enumerated assed
     /// Range 0 - 0x00000000FFFFFFFF (2^32)-1 is reserved for protected tokens
     /// the values under 1000 are used for ISO 4217 Numeric Curency codes
-    Asset(u64),
+    Asset(u128),
 }
 
 impl Default for AssetId {
